Extract Android platform lifecycle quirks from ScopedViewModelContainer to PlatformLifecycleHandler
