+++ resacacore/src/iosMain/kotlin/com/sebaslogen/resaca/core/PlatformLifecycleHandler.ios.kt
@@ -1,0 +1,15 @@
+package com.sebaslogen.resaca.core
+
+public actual class PlatformLifecycleHandler {
+    public actual fun onResumed() {
+        // No-op
+    }
+
+    public actual fun onDestroyed() {
+        // No-op
+    }
+
+    public actual suspend fun awaitBeforeDisposing(inForeground: Boolean) {
+        // No-op
+    }
+}
