+++ sample/src/androidTest/java/com/sebaslogen/resacaapp/sample/koin/ComposeActivityRecreationTests.kt
@@ -8,7 +8,7 @@ import androidx.test.core.app.ActivityScenario
 import androidx.test.core.app.ApplicationProvider
 import androidx.test.ext.junit.runners.AndroidJUnit4
 import androidx.test.filters.LargeTest
-import com.sebaslogen.resaca.COMPOSITION_RESUMED_TIMEOUT_IN_SECONDS
+import com.sebaslogen.resaca.core.COMPOSITION_RESUMED_TIMEOUT_IN_SECONDS
 import com.sebaslogen.resacaapp.sample.ui.main.ComposeActivity
 import com.sebaslogen.resacaapp.sample.ui.main.koinSingleViewModelScopedDestination
 import com.sebaslogen.resacaapp.sample.ui.main.koinViewModelScopedDestination
